+++ src/FactoredMatrices.jl
@@ -13,7 +13,7 @@ module FactoredMatrices
 import LinearAlgebra
 using LinearAlgebra: Adjoint, Bidiagonal, Diagonal, Factorization, Hermitian, LowerTriangular, SVD,
     SymTridiagonal, Symmetric, Transpose, Tridiagonal, UniformScaling, UnitLowerTriangular,
-    UnitUpperTriangular, UpperTriangular, dot, mul!, norm, qr, rank, svd, tr
+    UnitUpperTriangular, UpperTriangular, dot, mul!, norm, qr, rank, svd
 
 export FactoredMatrix
 public Workspace, CachedFactoredMatrix